backend/requirements.txt
flask
psycopg2-binary
kafka-python
flask-cors
